requirements.txt
slack_sdk
python-dotenv
norgatedata
sentry-sdk
selenium
psycopg2